Hasta grabación de pedidos y rutero

+++ PIGGISWS/Models/Politica.cs
@@ -8,7 +8,8 @@ public class Politica
     public decimal POL_PORC_FINANC { get; set; }
     public decimal POL_PORC_PRO_PAGO { get; set; }
     public decimal POL_PORC_PAG_CONTA { get; set; }
-    public decimal POL_LINEA_CREDITO { get; set; }
+    public int POL_LINEA_CREDITO { get; set; }
     public int POL_DIAS_PLAZO { get; set; }
     public int POL_NRO_PAGOS { get; set; }
+    
 }
